Added some analyzers for better indexing. Also continued on the auto-complete functionality.
git-svn-id: http://ikube.googlecode.com/svn/trunk@1003 89a0fb20-889e-aefa-90f6-7c296b20d5cf

+++ code/gui/src/test/java/ikube/listener/SnapshotListenerIntegration.java
@@ -22,7 +22,7 @@ public class SnapshotListenerIntegration extends Integration {
 	public void handleNotification() {
 		Event event = new Event();
 		event.setType(Event.PERFORMANCE);
-		long maxSnapshots = IConstants.MAX_SNAPSHOTS + 10;
+		double maxSnapshots = IConstants.MAX_SNAPSHOTS + 10d;
 		for (int i = 0; i < maxSnapshots; i++) {
 			snapshotListener.handleNotification(event);
 		}
